+++ .prettierignore
@@ -1,5 +1,8 @@
 back-end-py
 back-end-deployment-helper
+ 
 wrangler.jsonc
+dev.jsonc
+worker.d.ts
+
 front-end/index.min.html
-worker-configuration.d.ts
